fix(bun): fail the run when a global teardown throws
teardownSetups swallowed every teardown error while the exit code was
computed from file results alone, so a root could pass all its tests, fail
its cleanup contract, and still exit 0. Teardowns still all run in reverse
order; their failures now count. Two regressions cover it, and both fail
against the previous implementation.

Also corrects the PR body, which claimed repository-wide completion while
agents, cli and the SecureStore backend matrix still execute Vitest. Those
are enumerated explicitly instead.
