+++ scripts/tests/run_bun_tests.test.ts
@@ -596,130 +596,6 @@ describe('runBunTests', () => {
       'Native Bun test failed: /repo/e2e/broken.test.ts (exit code: 1)',
     );
   });
-
-  it('runs global setup before any child and teardown after the last one', async () => {
-    const order: string[] = [];
-    const dependencies: BunTestRunnerDependencies = {
-      repoRoot: '/repo',
-      invocationDirectory: '/invoke',
-      executable: '/bin/bun',
-      environment: {},
-      resolveFiles: () => [
-        {
-          cwd: '/repo/e2e',
-          file: '/repo/e2e/a.test.ts',
-          preloads: [],
-          globalSetup: '/repo/e2e/globalSetup.ts',
-        },
-      ],
-      resolveTsconfig: resolveTsconfigOverride,
-      spawn: () => {
-        order.push('spawn');
-        return { exitCode: 0, signalCode: null };
-      },
-      loadGlobalSetup: async () => ({
-        setup: () => {
-          order.push('setup');
-        },
-        teardown: () => {
-          order.push('teardown');
-        },
-      }),
-      stdout: () => {},
-      stderr: () => {},
-    };
-
-    await runBunTests([], dependencies);
-
-    expect(order).toEqual(['setup', 'spawn', 'teardown']);
-  });
-
-  it('tears down already-started setups when a later setup throws', async () => {
-    const order: string[] = [];
-    const dependencies: BunTestRunnerDependencies = {
-      repoRoot: '/repo',
-      invocationDirectory: '/invoke',
-      executable: '/bin/bun',
-      environment: {},
-      resolveFiles: () => [
-        {
-          cwd: '/repo/e2e',
-          file: '/repo/e2e/a.test.ts',
-          preloads: [],
-          globalSetup: '/repo/e2e/s1.ts',
-        },
-        {
-          cwd: '/repo/evals',
-          file: '/repo/evals/b.eval.ts',
-          preloads: [],
-          globalSetup: '/repo/evals/s2.ts',
-        },
-      ],
-      resolveTsconfig: resolveTsconfigOverride,
-      spawn: () => {
-        order.push('spawn');
-        return { exitCode: 0, signalCode: null };
-      },
-      loadGlobalSetup: async (path) =>
-        path === '/repo/e2e/s1.ts'
-          ? {
-              setup: () => {
-                order.push('setup1');
-              },
-              teardown: () => {
-                order.push('teardown1');
-              },
-            }
-          : {
-              setup: () => {
-                order.push('setup2');
-                throw new Error('setup boom');
-              },
-            },
-      stdout: () => {},
-      stderr: () => {},
-    };
-
-    await expect(runBunTests([], dependencies)).rejects.toThrow('setup boom');
-    // No file may run once setup failed, and the setup that did succeed must
-    // still be torn down or its resources leak.
-    expect(order).toEqual(['setup1', 'setup2', 'teardown1']);
-  });
-
-  it('runs global teardown even when a test file fails', async () => {
-    const order: string[] = [];
-    const dependencies: BunTestRunnerDependencies = {
-      repoRoot: '/repo',
-      invocationDirectory: '/invoke',
-      executable: '/bin/bun',
-      environment: {},
-      resolveFiles: () => [
-        {
-          cwd: '/repo/e2e',
-          file: '/repo/e2e/a.test.ts',
-          preloads: [],
-          globalSetup: '/repo/e2e/globalSetup.ts',
-        },
-      ],
-      resolveTsconfig: resolveTsconfigOverride,
-      spawn: () => ({ exitCode: 1, signalCode: null }),
-      loadGlobalSetup: async () => ({
-        setup: () => {
-          order.push('setup');
-        },
-        teardown: () => {
-          order.push('teardown');
-        },
-      }),
-      stdout: () => {},
-      stderr: () => {},
-    };
-
-    const status = await runBunTests([], dependencies);
-
-    expect(status).toBe(1);
-    expect(order).toEqual(['setup', 'teardown']);
-  });
 });
 
 describe('applyExclusions', () => {
